tracker/requirements.txt
websockets
clickhouse-driver
python-dotenv
shapely